add hills and get algo closer to real dijkstra
Still am not visiting cells that have been searched. Relying on first searched = shortest path and this only works because of the nature of the grid. Not sure if this would not work with random plots and vectors. Can't get it functioning the way I think it should work yet.